test_uv_append: Refactor out some utility macros
